% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{cordist}
\alias{cordist}
\title{Similarity measure which combines elements from Pearson correlation and
Euclidean distance.}
\usage{
cordist(dat)
}
\description{
Note that: dat is a data matrix with observations as columns (samples) and features as rows (genes)
}
