% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{export_network_to_graphml}
\alias{export_network_to_graphml}
\title{Converts an adjaceny matrix along with some optional vertex and edge
 information to a GraphML graph and saves it to disk.}
\usage{
export_network_to_graphml(adj_mat, filename = NULL, weighted = TRUE,
  threshold = 0.5, max_edge_ratio = 3, nodeAttr = NULL,
  nodeAttrDataFrame = NULL, edgeAttributes = NULL, verbose = FALSE)
}
\arguments{
\item{adj_mat}{An n-by-n weighted or unweighted adjacency matrix normalized
to contain values between 0 and 1.}

\item{filename}{Name of file to save output to. If file already exists it
will be overwritten. (default: network.graphml)}

\item{weighted}{Whether or not the adjacency matrix should be treated as a 
weighted graph. (default: TRUE)}

\item{threshold}{For weighted networks, if a threshold value between 0 and 
1 is specified, all edges with weights below that value with be dropped from
  the graph. (default: 0.5)}

\item{max_edge_ratio}{The maximum number of edges per node in the network to
allow. If the number of edges that would remain for the specified threshold
exceeds this value, the threshold will be raised to reduce the number of
edges remaining. (default: 3)}

\item{nodeAttr}{A vector with length equal to the number of vertices in the 
network, where the ith entry in the vector corresponds to some numeric or 
string annotation that should be associated with the ith node in the 
adjacency matrix. (default: NULL)}

\item{nodeAttrDataFrame}{A data frame containing one or more columns 
associated with the vertices in the graph.  The ith row of the dataframe 
should correspond to the ith entry in the adjacency matrix. (default: NULL)}

\item{edgeAttributes}{Extra attributes to associate with the graph edges,
formatted as a list of matrices of the same dimension and names as the
adjacency matrix.

Examples
--------
export_network_to_graphml(adj_mat, filename='~/network.graphml',
                          threshold=0.3, nodeAttrDataFrame=df)

See Also
--------
1. http://www.inside-r.org/packages/cran/WGCNA/docs/exportNetworkToCytoscape
2. http://graphml.graphdrawing.org/

Returns
-------
An igraph graph object representing the exported graph.}
}
\description{
Converts an adjaceny matrix along with some optional vertex and edge
 information to a GraphML graph and saves it to disk.
}
